tests/wpt/web-platform-tests/app-history/navigate/return-value/goTo-detach-same-document.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<iframe id="i" src="/common/blank.html"></iframe>

<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  assert_equals(i.contentWindow.appHistory.entries().length, 1);
  let key = i.contentWindow.appHistory.current.key;

  await i.contentWindow.appHistory.navigate("#1").committed;

  assert_equals(i.contentWindow.appHistory.entries().length, 2);
  assert_equals(i.contentWindow.appHistory.current, i.contentWindow.appHistory.entries()[1]);

  const iWindow = i.contentWindow;
  const iDOMException = iWindow.DOMException;
  i.contentWindow.appHistory.onnavigate = () => i.remove();

  await assertBothRejectDOM(t, i.contentWindow.appHistory.goTo(key), "AbortError", iWindow, iDOMException);
}, "goTo() promise rejections when detaching an iframe inside onnavigate (same-document)");
</script>
